Remove JSDoc
JSDoc is incapable of parsing the more advanced type annotations
recognized by the TypeScript compiler. Because there is no audience for
the documentation produced by JSDoc (this project is an application, not
a library) and because the recently-integrated TypeScript compiler
validates the JSDoc-style type annotations much more thoroughly, the
JSDoc package itself provides little value. Remove it.