Limit development task to development workflows
npm's "postinstall" script runs when a package is installed by consumers
(as in `npm install aria-at-automation-harness`). This makes it
unsuitable for tasks intended for local development (as in `husky
install`). Model the development task as a "prepare" script so that it
is only executed during local development workflows (as in `npm install`
from a local clone of this project's Git respository).